Extracting graph preparation from core_runner to an external graph_prep_runner
